NJ: swap back to old scrapers
